+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Fix all spec
 
 ### [Add]
+- Errors when no block given in transaction
 - Add timeout for transaction
 
 ## [2.7.2] - 2019-10-23